Elimino getters (no son necesarios)
